requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
async_cse==0.3.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
discord.py==2.5.2
dotenv==0.9.9
Flask==3.1.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.2
outcome==1.3.0.post0
pandas==2.3.1
propcache==0.3.2
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spotipy==2.25.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wavelink==3.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
